admin namespace and code refactoring
