+++ app/controllers/attachments_controller.rb
@@ -1,7 +1,6 @@
 class AttachmentsController < ApplicationController
   def create
-    @attachment = Attachment.new(attachment_params)
-    @attachment.save
+    @attachment = Attachment.create(attachment_params)
     render json: { attachment_id: @attachment.id, url: rails_blob_path(@attachment.file) }
   end
