+++ app/mailers/user_mailer.rb
@@ -4,7 +4,7 @@ class UserMailer < ApplicationMailer
     @user = User.find_by(id: recipient_id)
     mail(to: @user.email,
          # from: "info@publicv.org",
-         subject: 'Subject for mail',
+         subject: t('mailer.subject'),
          tag: "newsletter_#{newsletter_id}")
   end
 end
