+++ app/views/admins/newsletters/_preferences.html.erb
