+++ app/views/admins/newsletters/edit.html.erb
@@ -2,7 +2,7 @@
   <div class="pt-5">
     <div class="box box-primary mt-3 mb-5">
       <div class="box-header with-border">
-        <h3 class='box-title my-3 ml-4'><%= t('layouts.nav.new_newsletter') -%></h3>
+        <h3 class='box-title my-3 ml-4'><%= t('layouts.nav.edit_newsletter') -%></h3>
       </div>
       <div class="box-body">
         <%= render 'form' %>
