+++ app/views/admins/newsletters/index.html.erb
@@ -1,13 +1,13 @@
 <div class="container middle-container-1 cvpage">
   <div class="pt-5">
     <h1 class="text-primary ml-2"><%= t('layouts.nav.newsletter') %></h1>
-    <%= link_to new_newsletter_path, class:'text-decoration-none' do %>
+    <%= link_to new_admins_newsletter_path, class:'text-decoration-none' do %>
       <i class="fa fa-plus-circle ml-2"></i><%= t('layouts.nav.new_newsletter') %>
     <% end %>
     <div class="row mt-3">
       <div class="col-md-12">
         <div id="matters" class="box box-primary mb-4">
-          <%= form_tag newsletters_path, method: 'get' do %>
+          <%= form_tag admins_newsletters_path, method: 'get' do %>
             <table class="w-100 my-2">
               <thead>
                 <tr class="box-header">
@@ -28,15 +28,15 @@
                   <td><%= newsletter.recipient_ids %></td>
                   <td><%= newsletter.sent_at %></td>
                   <td>
-                    <%= link_to(edit_newsletter_path(newsletter),
+                    <%= link_to(edit_admins_newsletter_path(newsletter),
                                 title: 'Edit') do %>
                       <i class="fas fa-pencil-alt"></i>
                     <% end %>
-                    <%= link_to(newsletter_path(newsletter),
+                    <%= link_to(admins_newsletter_path(newsletter),
                                 title: 'Show', target: '_blank') do %>
                       <i class="fa fa-eye mx-1" aria-hidden="true"></i>
                     <% end %>
-                    <%= link_to(newsletter_submittal_index_path(newsletter),
+                    <%= link_to(admins_newsletter_submittal_index_path(newsletter),
                                 title: 'Send', method: :post, id: 'send_email') do %>
                       <i class="fa fa-paper-plane" aria-hidden="true"></i>
                     <% end %>
@@ -51,3 +51,4 @@
     </div>
   </div>
 </div>
+
