+++ app/views/layouts/shared/_header.html.erb
@@ -62,7 +62,7 @@
                   <% end %>
                   <% if (current_user.admin)%>
                     <div>
-                      <%= link_to t('layouts.nav.newsletter'), newsletters_path, class: 'dropdown-item py-2 badge-primary-lt' %>
+                      <%= link_to t('layouts.nav.newsletter'), admins_newsletters_path, class: 'dropdown-item py-2 badge-primary-lt' %>
                     </div>
                   <% else %>
                     <div>
