+++ config/locales/app.it.yml
@@ -163,6 +163,7 @@ it:
       presentation: 'il tuo CV Digitale'
       newsletter: 'Newsletter'
       new_newsletter: ' New Newsletter'
+      edit_newsletter: ' Edit Newsletter'
       preference: 'Preference'
     footer:
       links: 'Collegamenti utili'
