+++ config/locales/errors.en.yml
@@ -11,5 +11,6 @@ en:
       content_html: You are not authorized to access this page. <br>Possible reason for unauthorization could be your CV is not published.
       link: BACK TO HOME
     attributes:
-      name: 'Name: cannot be left blank'
-      recipient_ids: 'Recipient ids: Cannot be left blank'
+      name: 'cannot be left blank'
+      recipient_ids: 'cannot be left blank'
+    default_message: 'Cannot be saved due to %{count} errors:'
