+++ config/locales/mailers.en.yml
@@ -1,0 +1,3 @@
+en:
+  mailer:
+    subject: '(Subject for mail)'
